moved singleton to a separate module
