+++ singleton-service/Cargo.toml
@@ -15,3 +15,4 @@ log = "0.4"
 serde-humanize-rs = "0.1"
 proxy-wasm = "0.1"
 bincode = "1.0"
+chrono = "0.4.19"
