+++ singleton-service/src/lib.rs
@@ -1,3 +1,2 @@
 mod configuration;
 mod service;
-mod deltas;
